<OriginalCommitID:e9dd678> BUG fix  	[#5677] Problems with labeled statements
Made contine statements similar to brake statements so that they have the targeted label. Did not fix the existing getLabeledStatement because it screws up the visiting of the tree since it provides a reference to a (possibly) higher statement in the tree, and therefore can cause stack overflows when visiting. The get and set labeled statements should be deprecated in next version (they don't work anyways)


git-svn-id: file:///svnroot/spoon/trunk/spoon-core@2017 b1f50995-73fd-43b8-8ca2-fbc405a04e82