<OriginalCommitID:8c13e7e> supports creation of  spoon objects on the fly for classes in class path
with the new feature to inspect inside classes in the source class path
there was a new bug introduced: the URLClassLoader inspecting each
referenced class had two problems. firstly, a memory performance
issue: many URLClassLoaders were being created with each invocation,
but all of them using the same class path. secondly, they class loader
was closed after loading each class, but this could cause the program
to throw a NoClassDefFoundError (because the class is found, but it
can't be loaded because the loader is closed).

now, it is the Environment who gives support to load external classes
(but yet found on the source class path), providing its own ClassLoader