<OriginalCommitID:b9133a7> Supports annotations on type.
Creates a CtTypeAnnotation available by a CtTypeReference
to add an annotation on any types in a source code.
Can add an annotation on the type in CtNewClass, on the
type in a cast, on an exception in a thrown, on the return
type of a method, on the type of a local variable, on
generics types. Handles parameters for this type of
annotation too.